Try to refactor all methods
