+++ src/createSignedTx.spec.ts
@@ -1,11 +1,11 @@
-import { balanceTransfer } from './balanceTransfer';
 import { createSignedTx } from './createSignedTx';
 import { createSigningPayload } from './createSigningPayload';
+import { transfer } from './methods';
 import { signWithAlice, TEST_TRANSFER_TX_INFO } from './util/testUtil';
 
 describe('createSignedTx', () => {
   it('should work', async done => {
-    const unsigned = balanceTransfer(TEST_TRANSFER_TX_INFO);
+    const unsigned = transfer(TEST_TRANSFER_TX_INFO);
     const signingPayload = createSigningPayload(unsigned);
     const signature = await signWithAlice(signingPayload);
