+++ src/index.ts
@@ -1,10 +1,7 @@
-export * from './balanceTransfer';
 export * from './createSignedTx';
 export * from './createSigningPayload';
 export * from './decode/decode';
 export * from './deriveAddress';
 export * from './getTxHash';
 export * from './importPrivateKey';
-export * from './staking/bond';
-export * from './staking/nominate';
-export * from './staking/unbond';
+export * from './methods';
