+++ src/methods/index.ts
@@ -1,0 +1,2 @@
+export * from './balances';
+export * from './staking';
