+++ src/methods/staking/bond.spec.ts
