+++ src/methods/staking/index.ts
@@ -1,0 +1,3 @@
+export * from './bond';
+export * from './nominate';
+export * from './unbond';
