+++ src/methods/staking/nominate.ts
@@ -1,21 +1,30 @@
 import Metadata from '@polkadot/metadata';
 import { createType, TypeRegistry } from '@polkadot/types';
 
-import { EXTRINSIC_VERSION, ONE_SECOND } from '../util/constants';
-import { UnsignedTransaction } from '../util/types';
-import { TxInfoBond } from './stakingTxTypeUtils';
+import { EXTRINSIC_VERSION, ONE_SECOND } from '../../util/constants';
+import { BaseTxInfo, UnsignedTransaction } from '../../util/types';
+
+export interface TxInfoNominate extends BaseTxInfo {
+  /**
+   * The SS-58 encoded addresses of the targets you wish to nominate. A maximum of 16
+   * nominations are allowed.
+   *
+   * Warning: This provides no checks as to whether these targets are actual validators.
+   */
+  targets: Array<string>;
+}
 
 /**
- * Construct a transaction to bond funds and create a Stash account.
+ * Construct a transaction to nominate. This must be called by the _Controller_ account.
  *
  * @param info - Information required to construct the transaction.
  */
-export function bond(info: TxInfoBond): UnsignedTransaction {
+export function nominate(info: TxInfoNominate): UnsignedTransaction {
   const registry = new TypeRegistry();
   const metadata = new Metadata(registry, info.metadataRpc);
 
-  const bond = metadata.tx.staking.bond;
-  const method = bond(info.controller, info.value, info.payee).toHex();
+  const nominate = metadata.tx.staking.nominate;
+  const method = nominate(info.targets).toHex();
 
   return {
     address: info.address,
