+++ src/methods/staking/unbond.ts
@@ -1,21 +1,28 @@
 import Metadata from '@polkadot/metadata';
 import { createType, TypeRegistry } from '@polkadot/types';
 
-import { EXTRINSIC_VERSION, ONE_SECOND } from '../util/constants';
-import { UnsignedTransaction } from '../util/types';
-import { TxInfoNominate } from './stakingTxTypeUtils';
+import { EXTRINSIC_VERSION, ONE_SECOND } from '../../util/constants';
+import { BaseTxInfo, UnsignedTransaction } from '../../util/types';
+
+export interface TxInfoUnbond extends BaseTxInfo {
+  /**
+   * The number of tokens to unbond.
+   */
+  value: number;
+}
 
 /**
- * Construct a transaction to nominate. This must be called by the _Controller_ account.
+ * Construct a transaction to unbond funds from a Stash account. This must be called
+ * by the _Controller_ account.
  *
  * @param info - Information required to construct the transaction.
  */
-export function nominate(info: TxInfoNominate): UnsignedTransaction {
+export function unbond(info: TxInfoUnbond): UnsignedTransaction {
   const registry = new TypeRegistry();
   const metadata = new Metadata(registry, info.metadataRpc);
 
-  const nominate = metadata.tx.staking.nominate;
-  const method = nominate(info.targets).toHex();
+  const unbond = metadata.tx.staking.unbond;
+  const method = unbond(info.value).toHex();
 
   return {
     address: info.address,
